test: set up tests
Create one simple unit test for the login page and a snapshot test for the hero and about sections of the landing page.

Follow the convention where page tests are placed in the same directory as the respective `page.tsx` (like with login); snapshot tests and other necessary component tests are placed inside the `__tests__` folder.

Closes #91